requirements.txt
google-generativeai
pinecone-client
streamlit
grpcio
dotenv